backend/requirements.txt
Django>=3.0.8
psycopg2>=2.8.6
psycopg2-binary>=2.9.1
djangorestframework
django-cors-headers
